feat: migrate PersonView.php to MVC at /people/view/{personID}"
Agent-Logs-Url: https://github.com/ChurchCRM/CRM/sessions/d93d5ca4-a91c-4d89-9150-217573c9f3f5
