+++ src/people/index.php
@@ -14,5 +14,6 @@ require __DIR__ . '/routes/dashboard.php';
 require __DIR__ . '/routes/people.php';
 require __DIR__ . '/routes/family.php';
 require __DIR__ . '/routes/person.php';
+require __DIR__ . '/routes/view.php';
 
 $app->run();
